refactor: rename QuarkusPhysicalNamingStrategy
